+++ tools/linters/retrypolicy/analyzer_test.go
@@ -1,0 +1,12 @@
+package retrypolicy
+
+import (
+	"testing"
+
+	"golang.org/x/tools/go/analysis/analysistest"
+)
+
+func TestRetryPolicy(t *testing.T) {
+	testdata := analysistest.TestData()
+	analysistest.Run(t, testdata, Analyzer, "retrypolicy")
+}